requirements.txt
mysql-connector-python>=8.0.26 
configparser>=5.0.0 